Fix Stripe Connect "Failed to create Stripe connection" error
- Convert all Stripe Connect routes from .ts to .js per project guidelines
  (TypeScript inference issues with Stripe SDK types)
- Add granular error handling: missing API key, stale account recovery,
  account creation failures, and account link failures now return
  specific user-friendly error messages instead of one generic catch-all
- Handle stale Stripe Connect account IDs: if a stored account no longer
  exists in Stripe, clear it and create a fresh one instead of failing
